feature-remove all completed tasks
