--- src/clearCompleted.js
@@ -1,11 +1,0 @@
-// const ul = document.getElementById('list-items');
-// const li = document.getElementsByClassName('list-class');
-// const oldTasks = JSON.parse(localStorage.getItem('tasks'));
-
-// export const clearCompleted = () => {
-//   for(let i = 0; i < li.length; i++) {
-//     if(oldTasks[i].isCompleted == true){
-//       console.log(oldTasks[i]);
-//     }
-//   }
-// }
