+++ src/index.html
@@ -15,7 +15,9 @@
         <input type="text" placeholder="Add a Task" id="title" class="input-text">
         <i class="fa fa-sign-in"></i>
     </form>
-    <ul id="list-items"></ul>
-    <button type="button" class="submit">Clear</button>
+    <div id="div">
+        <ul id="list-items"></ul>
+    </div>
+    <button type="submit" id="submit">Clear</button>
 </body>
 </html>
